backend/requirements.txt
asgiref==3.6.0
boto3==1.26.59
botocore==1.29.59
Django==4.1.5
django-environ==0.9.0
django-ses==3.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
jmespath==1.0.1
mysqlclient==2.1.1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
